Hided features that are not working
