+++ src/components/common/Sidebar/SidebarLinks.utils.js
@@ -42,11 +42,20 @@ export const bottomSharedLinks = [
   },
 ];
 
+/**
+ * Khaleel Musleh
+ * Uncommented all the links which are non-functional, they are all available in the devLinks below
+ */
+
 export const menteeLinks = [
-  {
-    key: '/assignments',
-    label: 'My Assignments',
-  },
+  // {
+  //   key: '/assignments',
+  //   label: 'My Assignments',
+  // },
+  // {
+  //        key: '/meetings',
+  //       label: 'Meetings',
+  // },
   {
     key: '/resources',
     label: 'Access Resources',
@@ -58,10 +67,10 @@ export const menteeLinks = [
 ];
 
 export const mentorLinks = [
-  {
-    key: '/mentees',
-    label: 'My Mentees',
-  },
+  // {
+  //   key: '/mentees',
+  //   label: 'My Mentees',
+  // },
   {
     key: '/resources',
     label: 'Manage Resources',
@@ -70,6 +79,18 @@ export const mentorLinks = [
 ];
 
 export const adminLinks = [
+  // { key: '/dashboard', label: 'Dashboard', icon: <DashboardOutlined /> },
+  // { key: '/resources', label: 'Resources', icon: <BookOutlined /> },
+  { key: '/applications', label: 'Applications', icon: <ContainerOutlined /> },
+  { key: '/matching', label: 'Matching', icon: <UsergroupAddOutlined /> },
+  { key: '/users', label: 'Users', icon: <ContactsOutlined /> },
+  // { key: '/attendance', label: 'Attendance', icon: <FormOutlined /> },
+  // { key: '/meetings', label: 'Meetings', icon: <CarryOutOutlined /> },
+  // { key: '/reviews', label: 'Reviews' },
+  // { key: '/addMentorReview', label: 'Add Mentor Review' },
+];
+
+export const devLinks = [
   { key: '/dashboard', label: 'Dashboard', icon: <DashboardOutlined /> },
   { key: '/resources', label: 'Resources', icon: <BookOutlined /> },
   { key: '/applications', label: 'Applications', icon: <ContainerOutlined /> },
@@ -79,4 +100,6 @@ export const adminLinks = [
   { key: '/meetings', label: 'Meetings', icon: <CarryOutOutlined /> },
   { key: '/reviews', label: 'Reviews' },
   { key: '/addMentorReview', label: 'Add Mentor Review' },
+  { key: '/mentees', label: 'My Mentees' },
+  { key: '/assignments', label: 'My Assignments' },
 ];
